# Use memchr crate to speed up scanning specific chars