--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,6 +19,7 @@ doc = false
 [dependencies]
 bitflags = "1.0"
 unicase = "2.2.0"
+memchr = "2.2"
 getopts = { version = "0.2", optional = true }
 
 [dev-dependencies]