Used `memchr` to speed up two routines that scan for a single character and ran 11 benchmark loops on the crdt.md file.

desc  | memchr runtime (ns) | baseline runtime (ns) | thruput quotient
-- | --: | --: | --:
run 0  | 580920 | 590566 |  101.66%
run 1  | 565808 | 586689 |  103.69%
run 2  | 589241 | 591851 |  100.44%
run 3  | 568167 | 592621 |  104.30%
run 4  | 565312 | 590191 |  104.40%
run 5  | 584687 | 597149 |  102.13%
run 6  | 573395 | 590270 |  102.94%
run 7  | 566380 | 588118 |  103.84%
run 8  | 579796 | 595598 |  102.73%
run 9  | 569794 | 594025 |  104.25%
run 10  | 572270 | 589146 |  102.95%
**avg** | **631577** | **650622.4** | **103.02%**
**min** | **565312** | **586689** | **103.78%**